Updated swipe gesture for search.
